add-spending: fix bug with input null category; peroid-spending: refactor category loop logic in the circle-categories component & add update expense-service after save new expend
