+++ src/app/core/UI/calendar/calendar.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import { SelectorComponent } from './selector/selector.component';
 import moment from 'moment';
-import { DateService } from '../../../pages/spending/date.service';
+import { DateService } from '../../../service/date.service';
 import { DateFormatPipe } from './date-format.pipe';
