+++ src/app/pages/spending/components/add-spending/add-spending.component.scss
@@ -21,8 +21,6 @@
   padding-bottom: 10px;
   border-bottom: 1px solid #E8E8E8;
   
-
-
   ::ng-deep .mat-mdc-select-arrow {
     display: none;
   }
@@ -45,13 +43,10 @@
 }
 
 .name-of-product {
-  
   ::placeholder {
     font-size: 1rem;
   }
 }
-
-
 
 .choise-btn {
   display: flex;
