--- src/app/pages/spending/components/add-spending/add-spending.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { AddSpendingComponent } from './add-spending.component';
-
-describe('AddSpendingComponent', () => {
-  let component: AddSpendingComponent;
-  let fixture: ComponentFixture<AddSpendingComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [AddSpendingComponent]
-    })
-    .compileComponents();
-    
-    fixture = TestBed.createComponent(AddSpendingComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
