+++ src/app/pages/spending/components/add-spending/add-spending.component.ts
@@ -1,16 +1,16 @@
 import { ChangeDetectionStrategy, Component, Inject } from '@angular/core';
-import {MatSelectModule} from '@angular/material/select';
-import {MatFormFieldModule} from '@angular/material/form-field';
+import { MatSelectModule } from '@angular/material/select';
+import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatIconModule } from '@angular/material/icon';
-import {MatButtonModule} from '@angular/material/button';
-import {MatInputModule} from '@angular/material/input';
-import { ExpenseService } from '../../expense.service';
+import { MatButtonModule } from '@angular/material/button';
+import { MatInputModule } from '@angular/material/input';
+import { ExpenseService } from '../../../../service/expense.service';
 import { FormsModule } from '@angular/forms';
-import {MatNativeDateModule} from '@angular/material/core';
-import {MatDatepickerModule} from '@angular/material/datepicker';
-import moment from 'moment';
+import { MatNativeDateModule } from '@angular/material/core';
+import { MatDatepickerModule } from '@angular/material/datepicker';
 import { CATEGORY_SPENDING, ICategorySpending, ISpending } from '../../../../domain/spending.domain';
 import { MAT_BOTTOM_SHEET_DATA } from '@angular/material/bottom-sheet';
+import moment from 'moment';
 
 
 const UI_MODULES = [
@@ -43,16 +43,19 @@ export class AddSpendingComponent {
     private expenseService: ExpenseService,
     @Inject(MAT_BOTTOM_SHEET_DATA) public data: { category?: ICategorySpending }
   ) {
-    this.selectedCategory = this.data.category ? this.data.category : { title: '', icon: '' };
+    const isCategoryPreselected: boolean = !!(this.data && this.data.category);
+    const emptyCategory: ICategorySpending = { title: '', icon: '' };
+
+    this.selectedCategory = isCategoryPreselected ? this.data.category : emptyCategory;
   }
 
-  public saveSpending() {
+  public saveSpending(): void {
     const newExpense: ISpending = {
       icon: this.selectedCategory.icon,
       title: this.nameOfProduct,
       cost: this.costOfProduct,
       date: this.date,
-    }
+    }  
     this.expenseService.addSpending(newExpense);
   }
 }
