+++ src/app/pages/spending/components/period-spending/period-spending.component.html
@@ -7,14 +7,17 @@
         [color]="'#CBCACA'">
         </pgz-donut>
 
-        <a (click)="addSpending(categories[5])" class="item deg-1"><img src="assets/expend/health.svg"/></a>
-        <a (click)="addSpending(categories[7])" class="item deg-2"><img src="assets/expend/pet.svg"/></a>
-        <a (click)="addSpending(categories[0])" class="item deg-3"><img src="assets/expend/car.svg"/></a>
-        <a (click)="addSpending(categories[1])" class="item deg-4"><img src="assets/expend/clothes.svg"/></a>
-        <a (click)="addSpending(categories[2])" class="item deg-5"><img src="assets/expend/drink.svg"/></a>
-        <a (click)="addSpending(categories[4])" class="item deg-6"><img class="central-img" src="assets/expend/gift.svg"/></a>
-        <a (click)="addSpending(categories[3])" class="item deg-7"><img class="central-img" src="assets/expend/food.svg"/></a>
-        <a (click)="addSpending(categories[6])" class="item deg-8"><img src="assets/expend/house.svg" /></a>
+        @for(category of categories; track category; let i = $index) {
+            <a 
+            (click)="addSpending(category)"
+            class="item"
+            [ngClass]="'deg-' + (i + 1)">
+            @if(i === 5 || i === 6) {
+                <img [src]="category.icon" class="central-img"/>
+            } @else {
+                <img [src]="category.icon"/>
+            }
+         </a>
+        }
     </div>
-
 </section>
